moving unit-tests into dedicated tests/ dir
